fix: update audio processing to prevent float32 to int16 conversion warnings
- Modified audio generation functions across multiple tabs to clip audio values to the range [-1, 1] and convert to int16 format.
- Updated `load_realtime_model` to include `torch_dtype` parameter for codec loading.
- Ensured consistent audio output format for real-time inference, sound effects, TTS, and voice generation functionalities.

+++ app/model_loader.py
@@ -240,7 +240,9 @@ def load_realtime_model(device_str: str, attn_implementation: str):
 
     tokenizer = AutoTokenizer.from_pretrained(local_model_path)
 
-    codec = AutoModel.from_pretrained(local_codec_path, trust_remote_code=True).eval().to(device)
+    codec = AutoModel.from_pretrained(
+        local_codec_path, trust_remote_code=True, torch_dtype=dtype
+    ).eval().to(device)
 
     inferencer = MossTTSRealtimeInference(
         model, tokenizer,
